backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
 "aioodbc>=0.5.0",
 "azure-identity>=1.23.0",
 "azure-storage-blob>=12.25.1",
 "bcrypt>=4.3.0",
 "black>=25.1.0",
 "bleach>=6.2.0",
 "ebooklib>=0.19",
 "fastapi-users[sqlalchemy]>=14.0.1",
 "fastapi[standard]>=0.115.12",
 "flake8>=7.2.0",
 "isort>=6.0.1",
 "passlib>=1.7.4",
 "pillow>=11.2.1",
 "psycopg2-binary>=2.9.10",
 "pyodbc>=5.2.0",
 "python-jose>=3.4.0",
 "redis>=6.0.0",
 "sqlalchemy>=2.0.40",
]

[project.optional-dependencies]
dev = ["black>=24.10.0", "mypy>=1.14.1", "flake8>=7.1.1", "isort>=5.13.2"]
